Rename markdownImagePlugin to imagePlugin
