+++ src/build-scripts/build-database/imagePlugin.ts
